Moved joint file helpers into their own module to avoid problem where other classes were trying to use the name property of a class like this: object.class.ancestors.map {|c| c.name}
